Fix mapping PRF to hasher in the pluto helper
crypto_get_hasher() expected an IKEv2 integ algorithm, not a PRF.
