+++ include/pluto/spdb.h
@@ -157,14 +157,6 @@ extern struct db_sa *sa_copy_sa(struct db_sa *sa, int extra);
 extern struct db_sa *sa_copy_sa_first(struct db_sa *sa);
 extern struct db_sa *sa_merge_proposals(struct db_sa *a, struct db_sa *b);
 
-extern int v2tov1_encr(enum ikev2_trans_type_encr encr);
-extern int v2tov1_encr_child(enum ikev2_trans_type_encr encr);
-
-extern int v2tov1_integ(enum ikev2_trans_type_integ v2integ);
-extern int v2tov1_integ_child(enum ikev2_trans_type_integ v2integ);
-
-extern int v2tov1_prf(enum ikev2_trans_type_prf v2prf);
-
 extern bool extrapolate_v1_from_v2(struct db_sa *sadb, lset_t policy, enum phase1_role role);
 
 /* in spdb_struct.c */
